+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Changes:
 
 - Support for rejection of hanging RPC requests (30s timeout)
+- Add `.stats()` to provider interfaces with running counters
 - Update to latest Substrate, Kusama & Polkadot static metadata